fix: address code review feedback
- Correct _queue_for docstring: routing now uses put_nowait on an unbounded
  queue, so the old blocking back-pressure rationale no longer applies; a slow
  drain grows the queue in memory rather than stalling signal-cli's stdout.
- Dedupe the _start_server test helper into conftest (was duplicated verbatim
  in test_rpc.py and test_envelope_hardening.py).
- Extend the parse-guard comment to name _maybe_refresh_roster and emit_inbound
  as intentionally unguarded (their failures stay fatal); only parse_envelope
  is guarded.
